fix(sistemas): extraer lógica de complemento a 2 - Closes #643
