+++ tests/modulos/test_complemento_a_2.py
@@ -4,7 +4,7 @@ Tests unitarios para las funciones de complemento a 2.
 
 import pytest
 
-from escuadra.modulos.sistemas.herramienta_complemento_a_2 import (
+from escuadra.modulos.sistemas.complemento_a_2 import (
     complemento_a_decimal,
     decimal_a_complemento,
 )
